# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Security AI Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rules/import": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Api Import Rules",
        "description": "离线导入规则库 zip（需求2.2）。\n\n支持三种来源的 zip（自动识别）：\n  1. 本系统 rulepack：zip 内含 rulepack.json（{rules:[...]}）\n  2. NVD 导出：zip 内含 NVD API 导出的 json（{vulnerabilities:[...]}）\n  3. GitHub advisory-database：从 github/advisory-database 下载的 zip，\n     内含多个 GHSA-*.json（每个是一个 advisory）\n导入时用服务端签名密钥重新签名（agent 用同一密钥验签）。",
        "operationId": "api_import_rules_api_v1_rules_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_import_rules_api_v1_rules_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/sigma-rules/summary": {
      "get": {
        "tags": [
          "sigma-rules"
        ],
        "summary": "Sigma Summary",
        "description": "Import summary suitable for the console dashboard tab.\n\nReturns counts and the skipped-reason list. If no import has run\nyet, returns an empty placeholder so the UI can show a \"no rules\nimported\" state without a 404.",
        "operationId": "sigma_summary_api_v1_sigma_rules_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Sigma Summary Api V1 Sigma Rules Summary Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/sigma-rules": {
      "get": {
        "tags": [
          "sigma-rules"
        ],
        "summary": "List Sigma Rules",
        "description": "Filtered list of imported Sigma rules.\n\nServed straight from the manifest -- no DB, no walk of the rules\ndirectory on every request. If the manifest is missing we 404 so\nthe operator notices the import never ran.",
        "operationId": "list_sigma_rules_api_v1_sigma_rules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "process_creation / file_event / ...",
              "title": "Category"
            },
            "description": "process_creation / file_event / ..."
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "critical / high / medium / low / informational",
              "title": "Level"
            },
            "description": "critical / high / medium / low / informational"
          },
          {
            "name": "os",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "linux / macos / windows",
              "title": "Os"
            },
            "description": "linux / macos / windows"
          },
          {
            "name": "detector_supported",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detector Supported"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "substring match on rule_id or title",
              "title": "Q"
            },
            "description": "substring match on rule_id or title"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Sigma Rules Api V1 Sigma Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sigma-rules/import": {
      "post": {
        "tags": [
          "sigma-rules"
        ],
        "summary": "Import Sigma Dry Run",
        "description": "Run the importer against ``body.path`` and return the result.\n\nDoes NOT write the manifest or copy any .yml files. The operator\nis expected to use the CLI for the actual filesystem change so\nthe change is auditable on the host rather than over HTTP.",
        "operationId": "import_sigma_dry_run_api_v1_sigma_rules_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Import Sigma Dry Run Api V1 Sigma Rules Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/detect/rules": {
      "get": {
        "tags": [
          "detection"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_detect_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleSummary"
                  },
                  "type": "array",
                  "title": "Response List Rules Api V1 Detect Rules Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_import_rules_api_v1_rules_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "规则库 zip 压缩包"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_api_import_rules_api_v1_rules_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleSummary": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "title",
          "level",
          "product",
          "service",
          "category"
        ],
        "title": "RuleSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}